+++ framework/core/Models/Comment.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Comment object
+ *
+ * @package Brocooly
+ * @subpackage Brocket
+ * @since 1.11.0
+ */
+
+declare(strict_types=1);
+
+namespace Brocooly\Models;
+
+use Timber\Comment as TimberComment;
+
+class Comment extends TimberComment
+{
+
+}