+++ web/app/themes/brocooly/src/Hooks/DisableEmoji.php
@@ -20,7 +20,8 @@ class DisableEmoji implements Hookable
 		add_action( 'init', [ $this, 'disableEmoji' ] );
 	}
 
-	public function disableEmoji() {
+	public function disableEmoji()
+	{
 		remove_action( 'admin_print_styles', 'print_emoji_styles' );
 		remove_action( 'wp_head', 'print_emoji_detection_script', 7 );
 		remove_action( 'admin_print_scripts', 'print_emoji_detection_script' );
@@ -34,7 +35,8 @@ class DisableEmoji implements Hookable
 		add_filter( 'emoji_svg_url', '__return_false' );
 	}
 
-	public function disableEmojiFromTinyMCE( $plugins ) {
+	public function disableEmojiFromTinyMCE( $plugins )
+	{
 		if ( is_array( $plugins ) ) {
 			return array_diff( $plugins, [ 'wpemoji' ] );
 		}