+++ web/app/themes/brocooly/src/Models/WP/Category.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Category model
+ *
+ * @package Brocooly
+ * @subpackage Brocket
+ * @since 1.11.0
+ */
+
+declare(strict_types=1);
+
+namespace Theme\Models\WP;
+
+use Brocooly\Models\Taxonomy;
+
+class Category extends Taxonomy
+{
+
+}