+++ web/app/themes/brocooly/src/Models/WP/Comment.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Base comment object
+ *
+ * @package Brocooly
+ * @subpackage Brocket
+ * @since 1.11.0
+ */
+
+declare(strict_types=1);
+
+namespace Theme\Models\WP;
+
+use Brocooly\Models\Comment as BaseComment;
+
+class Comment extends BaseComment
+{
+
+}