+++ web/app/themes/brocooly/src/Models/WP/User.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Base user object
+ *
+ * @package Brocooly
+ * @subpackage Brocket
+ * @since 1.11.0
+ */
+
+declare(strict_types=1);
+
+namespace Theme\Models\WP;
+
+use Brocooly\Models\User as BaseUser;
+
+class User extends BaseUser
+{
+
+}